# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QL Trading AI Admin API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: transfers
    description: Transfer requests
  - name: withdrawals
    description: Withdrawal requests
  - name: trades
    description: Trade records
  - name: dashboard
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentActivityResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailResponse"

  /users/{id}/balance:
    post:
      operationId: adjustUserBalance
      tags: [users]
      summary: Adjust user balance (add or remove)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustBalanceBody"
      responses:
        "200":
          description: Balance adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /users/{id}/status:
    post:
      operationId: updateUserStatus
      tags: [users]
      summary: Update user status (active, frozen, banned, temp_banned)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /transfers:
    get:
      operationId: listTransfers
      tags: [transfers]
      summary: List all transfer requests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Transfers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransfersListResponse"

  /transfers/{id}/approve:
    post:
      operationId: approveTransfer
      tags: [transfers]
      summary: Approve a transfer request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transfer approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /transfers/{id}/reject:
    post:
      operationId: rejectTransfer
      tags: [transfers]
      summary: Reject a transfer request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectBody"
      responses:
        "200":
          description: Transfer rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List all withdrawal requests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Withdrawals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalsListResponse"

  /withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      tags: [withdrawals]
      summary: Approve a withdrawal request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /withdrawals/{id}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [withdrawals]
      summary: Reject a withdrawal request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectBody"
      responses:
        "200":
          description: Withdrawal rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List all trades
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Trades list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradesListResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
        error:
          type: string
        message:
          type: string
      required: [ok, error]

    SuccessResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
      required: [ok]

    LoginBody:
      type: object
      properties:
        password:
          type: string
      required: [password]

    LoginResponse:
      type: object
      properties:
        ok:
          type: boolean
        token:
          type: string
      required: [ok, token]

    RejectBody:
      type: object
      properties:
        reason:
          type: string

    AdjustBalanceBody:
      type: object
      properties:
        amount:
          type: number
        operation:
          type: string
          enum: [add, subtract]
        note:
          type: string
      required: [amount, operation]

    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, frozen, banned, temp_banned]
        banDuration:
          type: string
          description: "Duration for temp ban: '1h', '1d', '7d'"
        reason:
          type: string
      required: [status]

    User:
      type: object
      properties:
        id:
          type: integer
        tg_id:
          type: integer
          nullable: true
        name:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        status:
          type: string
        balance:
          type: number
        sub_expires:
          type: string
          nullable: true
        ban_until:
          type: string
          nullable: true
        created_at:
          type: string
        verified:
          type: boolean
      required: [id, status, balance, verified]

    UserDetail:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            trades_count:
              type: integer
            withdrawals_count:
              type: integer
            transfers_count:
              type: integer

    Transfer:
      type: object
      properties:
        id:
          type: integer
        from_user_id:
          type: integer
        to_user_id:
          type: integer
        amount:
          type: number
        status:
          type: string
        reason:
          type: string
          nullable: true
        from_user_name:
          type: string
          nullable: true
        to_user_name:
          type: string
          nullable: true
        created_at:
          type: string
        processed_at:
          type: string
          nullable: true
      required: [id, from_user_id, to_user_id, amount, status]

    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        amount:
          type: number
        fee:
          type: number
        net_amount:
          type: number
        method:
          type: string
        address:
          type: string
          nullable: true
        status:
          type: string
        reason:
          type: string
          nullable: true
        user_name:
          type: string
          nullable: true
        created_at:
          type: string
        processed_at:
          type: string
          nullable: true
      required: [id, user_id, amount, fee, net_amount, method, status]

    Trade:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: string
          nullable: true
        symbol:
          type: string
        side:
          type: string
        amount:
          type: number
        entry_price:
          type: number
        tp:
          type: number
          nullable: true
        sl:
          type: number
          nullable: true
        status:
          type: string
        profit:
          type: number
          nullable: true
        opened_at:
          type: string
        closed_at:
          type: string
          nullable: true
      required: [id, user_id, symbol, side, amount, entry_price, status]

    DashboardStats:
      type: object
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        bannedUsers:
          type: integer
        totalBalance:
          type: number
        pendingWithdrawals:
          type: integer
        pendingWithdrawalsAmount:
          type: number
        pendingTransfers:
          type: integer
        openTrades:
          type: integer
        totalProfit:
          type: number
        todayRegistrations:
          type: integer
      required: [totalUsers, activeUsers, bannedUsers, totalBalance, pendingWithdrawals, pendingWithdrawalsAmount, pendingTransfers, openTrades, totalProfit, todayRegistrations]

    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        description:
          type: string
        user_name:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        status:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, type, description, created_at]

    UsersListResponse:
      type: object
      properties:
        ok:
          type: boolean
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required: [ok, users, total]

    UserDetailResponse:
      type: object
      properties:
        ok:
          type: boolean
        user:
          $ref: "#/components/schemas/UserDetail"
      required: [ok, user]

    TransfersListResponse:
      type: object
      properties:
        ok:
          type: boolean
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
        total:
          type: integer
      required: [ok, transfers, total]

    WithdrawalsListResponse:
      type: object
      properties:
        ok:
          type: boolean
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        total:
          type: integer
      required: [ok, withdrawals, total]

    TradesListResponse:
      type: object
      properties:
        ok:
          type: boolean
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
      required: [ok, trades, total]

    RecentActivityResponse:
      type: object
      properties:
        ok:
          type: boolean
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required: [ok, activities]