# lib/api-zod/src/generated/types/tradesListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QL Trading AI Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { Trade } from "./trade";

export interface TradesListResponse {
  ok: boolean;
  trades: Trade[];
  total: number;
}