# lib/api-zod/src/generated/types/transfer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QL Trading AI Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface Transfer {
  id: number;
  from_user_id: number;
  to_user_id: number;
  amount: number;
  status: string;
  reason?: string | null;
  from_user_name?: string | null;
  to_user_name?: string | null;
  created_at?: string;
  processed_at?: string | null;
}