# lib/api-zod/src/generated/types/transfersListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QL Trading AI Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { Transfer } from "./transfer";

export interface TransfersListResponse {
  ok: boolean;
  transfers: Transfer[];
  total: number;
}