# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QL Trading AI Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  tg_id?: number | null;
  name?: string | null;
  username?: string | null;
  email?: string | null;
  status: string;
  balance: number;
  sub_expires?: string | null;
  ban_until?: string | null;
  created_at?: string;
  verified: boolean;
}